ajustes A*:

+++ src/main/java/domain/NodeComparator.java
@@ -9,6 +9,6 @@ public class NodeComparator implements Comparator<Node> {
 
     @Override
     public int compare(Node node1, Node node2) {
-        return Integer.compare(node1.getTotalCost(), node2.getTotalCost());
+        return Double.compare(node1.getTotalCost(), node2.getTotalCost());
     }
 }
